接入添加Oracle-cdc类型

+++ fisk-factory-access/fisk-factory-access-entity/src/main/java/com/fisk/dataaccess/dto/oraclecdc/CdcJobParameterDTO.java
@@ -15,7 +15,7 @@ public class CdcJobParameterDTO {
     /**
      * 目标表id
      */
-    public Long tableAccessId;
+    public Integer tableAccessId;
 
     /**
      * 目标表名称
